requirements.txt
# Non RobotPy packages that we use
black
pylint==2.15.3
pytest
wheel
# RobotPy packages.  We explicitly list the version number on all these packages so that when RobotPy
# updates we don't break the world until we have a chance to test/update.
pyfrc==2022.1.5
pynetconsole==2.0.4
pynetworktables==2021.0.0
pyntcore==2022.4.1.0
robotpy-commands-v2==2022.4.1.1
robotpy-ctre==2022.1.1
robotpy-hal==2022.4.1.2
robotpy-halsim-gui==2022.4.1.1
robotpy-installer==2022.1.1
robotpy-navx==2022.0.2
robotpy-wpimath==2022.4.1.3
robotpy-wpiutil==2022.4.1.2
robotpy-pathplannerlib==2022.2.3
wpilib==2022.4.1.6

